requirements.txt
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
transformers==4.43.3
torch>=2.1.0
sentence-transformers==3.0.1
scikit-learn==1.4.1.post1
tweepy==4.14.0
praw==7.7.1
facebook-sdk==3.1.0
tiktokapipy==0.2.5
hdbscan==0.8.36
streamlit==1.36.0
